Share outbox persist so deferred jobs use the same retry accounting.
